requirements.txt
asgiref==3.11.0
Django==6.0.1
django-debug-toolbar==6.1.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
